Creating Mongo connection,  User schema, and a POST request for User
First I connected a MonogDB instance to our Express server through Mongoose. Then, created a Mongoose schema for User and lastly created a POST request for User. Now the server can take in a POST to /user and creates a new User in the DB and returns a success or failure message.