Created front end API functions.
Added functions in the front end which can be called to access the backend API.
Add the statement 'import * as api from "{path}"' at the top of a file and you can then use any and all functions in the api.js file in the client folder. We will continue adding to these as we add more routes in the backend. Currently only supports POST and GET for User and Restaurant.