Login and logout now dynamically displays user name in header.
Data fetched from database in the header component and props are passed to all pages to interface with the login/logout cookie information. Login and logout now forces a page refresh (manually, in vanilla JS instead of with a state update (since there doesn't need to be a state variable to keep track of currently signed in session)). The page refresh accomplishes two things: clears the login field upon a successful login, and displays the name of the logged in user.

Will add a message indicating login failure next.