Star rating now included in reviews.
Note that an empty review (no review typed into the review box) results in a POST error. This, I guess, is not a problem since we don't want to be recording blank reviews in our database (I guess that means we don't support the option of just leaving a star rating without leaving context in the text box).