Reviews and ratings now live.
Reviews and ratings are now visible on the frontend. Data persists through multiple sessions, refreshes, etc. Everything is saved and fetched to and from the database.

Note that there's a level of indirection in the reviews backend. Specifically, review object ID's are saved in a restaurant object's reviews field, and this ID is used to find the corresponding review object with the description (actual review) and rating (star rating out of 5). The searching process is performed on the frontend. It's not a significant enough operation to noticeably impact performance of our app, not that performance is even considered in this project's grading.