First iteration of meaningful search done.
Can now search by cuisine type. The search box takes in a string input and outputs a list of all restaurants that have a cuisine that includes whatever string is inputted into the text box (case insensitive).

Next iteration of search: directly integrate searchbar in the RestaurantsPage componnet so that we can render only the buttons for restaurants that match the search (and no longer output a list of matching restaurants).